Removed IDisposable from IBlockchainProcessingStrategy. MinBlockConfirmations is not uint instead of ulong

+++ Nethereum.BlockchainProcessing/Processing/BlockchainSourceConfiguration.cs
@@ -11,7 +11,7 @@
         public string BlockchainUrl { get; set; }
         public string Name { get; set; }
         public ulong? MinimumBlockNumber { get; set; }
-        public ulong? MinimumBlockConfirmations { get; set; }
+        public uint? MinimumBlockConfirmations { get; set; }
         public ulong? FromBlock { get; set; }
         public ulong? ToBlock { get; set; }
         public bool PostVm { get; set; } = false;
